+++ code/src/App.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Thoughts } from './Thoughts'
+import { Thoughts } from './components/Thoughts'
 
 export const App = () => {
   return (
