+++ code/src/components/NewThought.js
