+++ code/src/components/NewThoughtForm.js
