+++ code/src/components/ThoughtCard.js
