+++ code/src/components/Thoughts.js
