+++ code/src/components/newthought.css
