+++ code/src/components/newthoughtform.css
