+++ code/src/components/thoughtcard.css
