+++ code/src/components/thoughts.css
